refactor: rename all Bud test files to TenantConfig
