+++ tests/Unit/Overrides/Auth/TenantConfigAuthManagerCreatorTest.php
@@ -1,14 +1,13 @@
 <?php
 declare(strict_types=1);
 
-namespace Sprout\Tests\Unit\Overrides\Broadcast;
+namespace Sprout\Tests\Unit\Overrides\Auth;
 
-use Illuminate\Contracts\Broadcasting\Broadcaster;
+use Illuminate\Contracts\Auth\UserProvider;
 use Illuminate\Foundation\Application;
-use InvalidArgumentException;
 use Mockery;
 use PHPUnit\Framework\Attributes\Test;
-use Sprout\Bud;
+use Sprout\TenantConfig;
 use Sprout\Contracts\ConfigStore;
 use Sprout\Contracts\Tenancy;
 use Sprout\Contracts\Tenant;
@@ -16,13 +15,13 @@ use Sprout\Contracts\TenantHasResources;
 use Sprout\Exceptions\TenancyMissingException;
 use Sprout\Exceptions\TenantMissingException;
 use Sprout\Managers\ConfigStoreManager;
-use Sprout\Overrides\Broadcast\BudBroadcastConnectionCreator;
-use Sprout\Overrides\Broadcast\BudBroadcastManager;
+use Sprout\Overrides\Auth\TenantConfigAuthManager;
+use Sprout\Overrides\Auth\TenantConfigAuthProviderCreator;
 use Sprout\Sprout;
 use Sprout\Support\SettingsRepository;
 use Sprout\Tests\Unit\UnitTestCase;
 
-class BudBroadcastConnectionCreatorTest extends UnitTestCase
+class TenantConfigAuthManagerCreatorTest extends UnitTestCase
 {
     private function mockApplication(bool $default = false): Application&Mockery\MockInterface
     {
@@ -31,17 +30,15 @@ class BudBroadcastConnectionCreatorTest extends UnitTestCase
         });
     }
 
-    private function mockManager(bool $driver = true): BudBroadcastManager&Mockery\MockInterface
+    private function mockManager(bool $driver = true): TenantConfigAuthManager&Mockery\MockInterface
     {
-        return Mockery::mock(BudBroadcastManager::class, static function (Mockery\MockInterface $mock) use ($driver) {
+        return Mockery::mock(TenantConfigAuthManager::class, static function (Mockery\MockInterface $mock) use ($driver) {
             if ($driver) {
-                $mock->shouldReceive('connectUsing')
+                $mock->shouldReceive('createUserProviderFromConfig')
                      ->with(
-                         'fake-connection',
-                         ['name' => 'fake-connection', 'driver' => 'null'],
-                         true
+                         ['provider' => 'fake-provider', 'driver' => 'null'],
                      )
-                     ->andReturn(Mockery::mock(Broadcaster::class))
+                     ->andReturn(Mockery::mock(UserProvider::class))
                      ->once();
             }
         });
@@ -58,8 +55,8 @@ class BudBroadcastConnectionCreatorTest extends UnitTestCase
                               ->with(
                                   $tenancy,
                                   $tenant,
-                                  'broadcast',
-                                  'fake-connection'
+                                  'auth',
+                                  'fake-provider'
                               )
                               ->andReturn([
                                   'driver' => 'null',
@@ -101,9 +98,9 @@ class BudBroadcastConnectionCreatorTest extends UnitTestCase
         return $sprout;
     }
 
-    private function getBud(Application $app, ConfigStoreManager $manager): Bud
+    private function getTenantConfig(Application $app, ConfigStoreManager $manager): TenantConfig
     {
-        return new Bud($app, $manager);
+        return new TenantConfig($app, $manager);
     }
 
     #[Test]
@@ -112,44 +109,19 @@ class BudBroadcastConnectionCreatorTest extends UnitTestCase
         $app     = $this->mockApplication();
         $manager = $this->mockManager();
         $config  = [
-            'name'   => 'fake-connection',
-            'driver' => 'fake-driver',
+            'provider' => 'fake-provider',
+            'driver'   => 'fake-driver',
         ];
         $sprout  = $this->getSprout($app);
-        $bud     = $this->getBud($app, $this->mockConfigStoreManager($sprout->getCurrentTenancy(), $sprout->getCurrentTenancy()->tenant()));
+        $tenantConfig     = $this->getTenantConfig($app, $this->mockConfigStoreManager($sprout->getCurrentTenancy(), $sprout->getCurrentTenancy()->tenant()));
 
-        $creator = new BudBroadcastConnectionCreator(
+        $creator = new TenantConfigAuthProviderCreator(
             $manager,
             $bud,
             $sprout,
+            'fake-provider',
             $config,
         );
-
-        $creator();
-    }
-
-    #[Test]
-    public function throwsAnExceptionWhenConfigIsMissingName(): void
-    {
-        $app     = $this->mockApplication();
-        $manager = $this->mockManager(false);
-        $config  = [];
-        $sprout  = $this->getSprout($app, false, false);
-        $bud     = $this->getBud($app, $this->mockConfigStoreManager());
-
-        $sprout->markAsOutsideContext();
-
-        $this->assertFalse($sprout->withinContext());
-
-        $creator = new BudBroadcastConnectionCreator(
-            $manager,
-            $bud,
-            $sprout,
-            $config
-        );
-
-        $this->expectException(InvalidArgumentException::class);
-        $this->expectExceptionMessage('Broadcast connection name must be provided');
 
         $creator();
     }
@@ -160,19 +132,20 @@ class BudBroadcastConnectionCreatorTest extends UnitTestCase
         $app     = $this->mockApplication();
         $manager = $this->mockManager(false);
         $config  = [
-            'name' => 'fake-connection',
+            'provider' => 'fake-provider',
         ];
         $sprout  = $this->getSprout($app, false, false);
-        $bud     = $this->getBud($app, $this->mockConfigStoreManager());
+        $tenantConfig     = $this->getTenantConfig($app, $this->mockConfigStoreManager());
 
         $sprout->markAsOutsideContext();
 
         $this->assertFalse($sprout->withinContext());
 
-        $creator = new BudBroadcastConnectionCreator(
+        $creator = new TenantConfigAuthProviderCreator(
             $manager,
             $bud,
             $sprout,
+            'fake-provider',
             $config
         );
 
@@ -188,19 +161,20 @@ class BudBroadcastConnectionCreatorTest extends UnitTestCase
         $app     = $this->mockApplication();
         $manager = $this->mockManager(false);
         $config  = [
-            'name' => 'fake-connection',
+            'provider' => 'fake-provider',
         ];
         $sprout  = $this->getSprout($app, false, false);
-        $bud     = $this->getBud($app, $this->mockConfigStoreManager());
+        $tenantConfig     = $this->getTenantConfig($app, $this->mockConfigStoreManager());
 
         $sprout->markAsInContext();
 
         $this->assertTrue($sprout->withinContext());
 
-        $creator = new BudBroadcastConnectionCreator(
+        $creator = new TenantConfigAuthProviderCreator(
             $manager,
             $bud,
             $sprout,
+            'fake-provider',
             $config
         );
 
@@ -216,17 +190,18 @@ class BudBroadcastConnectionCreatorTest extends UnitTestCase
         $app     = $this->mockApplication();
         $manager = $this->mockManager(false);
         $config  = [
-            'name' => 'fake-connection',
+            'provider' => 'fake-provider',
         ];
         $sprout  = $this->getSprout($app, true, false);
-        $bud     = $this->getBud($app, $this->mockConfigStoreManager());
+        $tenantConfig     = $this->getTenantConfig($app, $this->mockConfigStoreManager());
 
         $this->assertTrue($sprout->withinContext());
 
-        $creator = new BudBroadcastConnectionCreator(
+        $creator = new TenantConfigAuthProviderCreator(
             $manager,
             $bud,
             $sprout,
+            'fake-provider',
             $config
         );
