+++ tests/Unit/Overrides/Database/TenantConfigDatabaseConnectionOverrideTest.php
@@ -11,7 +11,7 @@ use Mockery;
 use Mockery\MockInterface;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Test;
-use Sprout\Bud;
+use Sprout\TenantConfig;
 use Sprout\Contracts\BootableServiceOverride;
 use Sprout\Contracts\ConfigStore;
 use Sprout\Contracts\Tenancy;
@@ -19,13 +19,13 @@ use Sprout\Contracts\Tenant;
 use Sprout\Contracts\TenantHasResources;
 use Sprout\Exceptions\CyclicOverrideException;
 use Sprout\Managers\ConfigStoreManager;
-use Sprout\Overrides\Database\BudDatabaseConnectionOverride;
+use Sprout\Overrides\Database\TenantConfigDatabaseConnectionOverride;
 use Sprout\Sprout;
 use Sprout\Support\SettingsRepository;
 use Sprout\Tests\Unit\UnitTestCase;
 use function Sprout\sprout;
 
-class BudDatabaseConnectionOverrideTest extends UnitTestCase
+class TenantConfigDatabaseConnectionOverrideTest extends UnitTestCase
 {
     protected function defineEnvironment($app): void
     {
@@ -39,7 +39,7 @@ class BudDatabaseConnectionOverrideTest extends UnitTestCase
         return Mockery::mock(DatabaseManager::class, static function (MockInterface $mock) use ($extends, $callback) {
             if ($extends) {
                 $mock->shouldReceive('extend')
-                     ->with('sprout:bud', Mockery::on(static function ($arg) {
+                     ->with('sprout:config', Mockery::on(static function ($arg) {
                          return is_callable($arg) && $arg instanceof Closure;
                      }))
                      ->once();
@@ -54,7 +54,7 @@ class BudDatabaseConnectionOverrideTest extends UnitTestCase
     #[Test]
     public function isBuiltCorrectly(): void
     {
-        $this->assertTrue(is_subclass_of(BudDatabaseConnectionOverride::class, BootableServiceOverride::class));
+        $this->assertTrue(is_subclass_of(TenantConfigDatabaseConnectionOverride::class, BootableServiceOverride::class));
     }
 
     #[Test]
@@ -64,7 +64,7 @@ class BudDatabaseConnectionOverrideTest extends UnitTestCase
 
         config()->set('sprout.overrides', [
             'database' => [
-                'driver' => BudDatabaseConnectionOverride::class,
+                'driver' => TenantConfigDatabaseConnectionOverride::class,
             ],
         ]);
 
@@ -73,7 +73,7 @@ class BudDatabaseConnectionOverrideTest extends UnitTestCase
         $sprout->overrides()->registerOverrides();
 
         $this->assertTrue($sprout->overrides()->hasOverride('database'));
-        $this->assertSame(BudDatabaseConnectionOverride::class, $sprout->overrides()->getOverrideClass('database'));
+        $this->assertSame(TenantConfigDatabaseConnectionOverride::class, $sprout->overrides()->getOverrideClass('database'));
         $this->assertTrue($sprout->overrides()->isOverrideBootable('database'));
         $this->assertTrue($sprout->overrides()->hasOverrideBooted('database'));
     }
@@ -81,7 +81,7 @@ class BudDatabaseConnectionOverrideTest extends UnitTestCase
     #[Test, DataProvider('managerResolvedDataProvider')]
     public function bootsCorrectly(bool $return): void
     {
-        $override = new BudDatabaseConnectionOverride('database', []);
+        $override = new TenantConfigDatabaseConnectionOverride('database', []);
 
         /** @var \Illuminate\Foundation\Application&MockInterface $app */
         $app = Mockery::mock($this->app, function (MockInterface $mock) use ($return) {
@@ -122,9 +122,9 @@ class BudDatabaseConnectionOverrideTest extends UnitTestCase
     }
 
     #[Test]
-    public function errorsIfOverriddenConnectionAlsoUsesBud(): void
+    public function errorsIfOverriddenConnectionAlsoUsesConfig(): void
     {
-        $override = new BudDatabaseConnectionOverride('database', []);
+        $override = new TenantConfigDatabaseConnectionOverride('database', []);
 
         $tenant  = Mockery::mock(Tenant::class, TenantHasResources::class, static function (MockInterface $mock) {
         });
@@ -140,11 +140,11 @@ class BudDatabaseConnectionOverrideTest extends UnitTestCase
 
         $app->make('config')->set('multitenancy.defaults.config', 'filesystem');
         $app->make('config')->set('database.connections.bud-connection', [
-            'driver' => 'sprout:bud',
+            'driver' => 'sprout:config',
             'database' => 'bud-database',
         ]);
 
-        $app->singleton(Bud::class, fn () => new Bud($app, Mockery::mock(ConfigStoreManager::class, function (MockInterface $mock) use ($tenancy, $tenant) {
+        $app->singleton(TenantConfig::class, fn () => new TenantConfig($app, Mockery::mock(ConfigStoreManager::class, function (MockInterface $mock) use ($tenancy, $tenant) {
             $mock->shouldReceive('get')
                  ->andReturn(Mockery::mock(ConfigStore::class, static function (MockInterface $mock) use ($tenancy, $tenant) {
                      $mock->shouldReceive('get')
@@ -154,7 +154,7 @@ class BudDatabaseConnectionOverrideTest extends UnitTestCase
                               'database',
                               'bud-connection',
                           )->andReturn([
-                             'driver' => 'sprout:bud',
+                             'driver' => 'sprout:config',
                              'database' => 'bud-database',
                          ]);
                  }));
@@ -178,9 +178,9 @@ class BudDatabaseConnectionOverrideTest extends UnitTestCase
     }
 
     #[Test]
-    public function keepsTrackOfResolvedBudDrivers(): void
+    public function keepsTrackOfResolvedConfigDrivers(): void
     {
-        $override = new BudDatabaseConnectionOverride('database', []);
+        $override = new TenantConfigDatabaseConnectionOverride('database', []);
 
         $tenant  = Mockery::mock(Tenant::class, TenantHasResources::class, static function (MockInterface $mock) {
         });
@@ -196,11 +196,11 @@ class BudDatabaseConnectionOverrideTest extends UnitTestCase
 
         $app->make('config')->set('multitenancy.defaults.config', 'filesystem');
         $app->make('config')->set('database.connections.bud-connection', [
-            'driver' => 'sprout:bud',
+            'driver' => 'sprout:config',
             'database' => 'bud-database',
         ]);
 
-        $app->singleton(Bud::class, fn () => new Bud($app, Mockery::mock(ConfigStoreManager::class, function (MockInterface $mock) use ($tenancy, $tenant) {
+        $app->singleton(TenantConfig::class, fn () => new TenantConfig($app, Mockery::mock(ConfigStoreManager::class, function (MockInterface $mock) use ($tenancy, $tenant) {
             $mock->shouldReceive('get')
                  ->andReturn(Mockery::mock(ConfigStore::class, function (MockInterface $mock) use ($tenancy, $tenant) {
                      $mock->shouldReceive('get')
@@ -235,7 +235,7 @@ class BudDatabaseConnectionOverrideTest extends UnitTestCase
     #[Test]
     public function cleansUpResolvedDrivers(): void
     {
-        $override = new BudDatabaseConnectionOverride('database', []);
+        $override = new TenantConfigDatabaseConnectionOverride('database', []);
 
         $this->app->forgetInstance('db');
 
@@ -246,7 +246,7 @@ class BudDatabaseConnectionOverrideTest extends UnitTestCase
 
         $app->make('config')->set('multitenancy.defaults.config', 'filesystem');
         $app->make('config')->set('database.connections.bud-connection', [
-            'driver' => 'sprout:bud',
+            'driver' => 'sprout:config',
             'database' => 'bud-database',
         ]);
 
@@ -260,7 +260,7 @@ class BudDatabaseConnectionOverrideTest extends UnitTestCase
             $mock->shouldReceive('tenant')->andReturn($tenant)->once();
         });
 
-        $app->singleton(Bud::class, fn () => new Bud($app, Mockery::mock(ConfigStoreManager::class, function (MockInterface $mock) use ($tenancy, $tenant) {
+        $app->singleton(TenantConfig::class, fn () => new TenantConfig($app, Mockery::mock(ConfigStoreManager::class, function (MockInterface $mock) use ($tenancy, $tenant) {
             $mock->shouldReceive('get')
                  ->andReturn(Mockery::mock(ConfigStore::class, function (MockInterface $mock) use ($tenancy, $tenant) {
                      $mock->shouldReceive('get')
@@ -298,7 +298,7 @@ class BudDatabaseConnectionOverrideTest extends UnitTestCase
     #[Test]
     public function cleansUpNothingWithoutResolvedDrivers(): void
     {
-        $override = new BudDatabaseConnectionOverride('database', []);
+        $override = new TenantConfigDatabaseConnectionOverride('database', []);
 
         $this->app->forgetInstance('db');
 
@@ -309,7 +309,7 @@ class BudDatabaseConnectionOverrideTest extends UnitTestCase
 
         $app->make('config')->set('multitenancy.defaults.config', 'filesystem');
         $app->make('config')->set('database.connections.bud-connection', [
-            'driver' => 'sprout:bud',
+            'driver' => 'sprout:config',
             'database' => 'bud-database',
         ]);
